Minor: Remove the `tmp` directory from post gen hook
Remove the `tmp` directory from the list of empty directories
that would be created by the post gen hook. The main purpose of
the `tmp` directory was to contain binaries for tests and tools,
such as GolangCI-Lint

The current setup of the project has been modified to ensure that
the directory won't be required. Binaries can be installed
normally to the default Go path, from where they'll be used