Update: Fetch Go version to use as a setup argument
Instead of simply hard-coding the Go version to use as `1.16`,
ask the user to chose one as per their requirement

The offered versions are `1.16` and `1.17` - i.e. the last two
major revisions. Users would be selecting an option from the
choices offered - same as with chosing a license file

The version chosen will be applied to the `go.mod` file as well as
GolangCI linter configuration settings for individual linters