Update: Changes to dockerfile to generate images
Update the dockerfile to use the version of Go chosen by the user
during setup.

Switch to using the stable version of Go docker image as the base
for the first image (in the multi-stage build). Change the name
of the binary generated inside docker container to use the project
name provided by the user during setup