Update: Replace instances of `go-template` used as placeholder
The project still had a few places where the name of the repository
i.e. `go-template` was being used instead of the name of the
project template generated by users on their end.

This commit fixes this issue, replacing such instances of
`go-template` with the name of the project chosen by users during
setup