Fix: Modify default value for module path in Cookiecutter configs
Until now, the expression being used to set default value for
`cookiecutter.license_owner` field would assume that the module
name belong to a Github repository. The default value for the
field would split the module name with this assumption in mind!
Here is the expression being used to populate the license owner
field by default:

      `{{ cookiecutter.go_module_path.split('/')[1] }}`

With the previous expression, if the module path entered was not
a link, and was a simple name - such as `local` - the previous
logic used to split string would result in a runtime crash

This commit modifies the previous behaviour, setting a hard-coded
default value for the `license_owner` field, ensuring this bug
won't occur again