Major: Large scale refactor, add inner README to the project
Refactor large parts of the project, this includes
  - Remove custom label from PRs raised by `depandabot`
  - Modify section titles for `Release Drafter`
  - Patch minor bug in `tests.yml` causing crashes
  - Changes to `Makefile` to improve aesthetics
  - Additional minor corrections in `docker/README.md`

Update the main `README` for the generated project. This includes
adding instructions on how to clone/setup the project, installing
all the tools necessary to get the project running. Guide on the
usage of the Makefile and the important commands present in Makefile.
Replace static links with dynamic links where possible.

Add a section with `Additional Resources` that will detail on
setting up Codecov integration, usage of `Release Drafter`, and the
labels it will use to group merged PRs and more!